requirements.txt
boto3
python-dateutil
pymysql
boto3
pylint
pytest
pytest-cov
pytest-env
black
autopep8
